Move span definitions from wrapper.py to entities
