+++ mlflow/entities/__init__.py
@@ -20,14 +20,13 @@ from mlflow.entities.run_inputs import RunInputs
 from mlflow.entities.run_status import RunStatus
 from mlflow.entities.run_tag import RunTag
 from mlflow.entities.source_type import SourceType
-from mlflow.entities.span import SpanType
+from mlflow.entities.span import NoOpSpan, LiveSpan, Span, SpanType
 from mlflow.entities.span_event import SpanEvent
 from mlflow.entities.span_status import SpanStatus, SpanStatusCode
 from mlflow.entities.trace import Trace
 from mlflow.entities.trace_data import TraceData
 from mlflow.entities.trace_info import TraceInfo
 from mlflow.entities.view_type import ViewType
-from mlflow.tracing.types.wrapper import LiveSpan, Span
 
 __all__ = [
     "Experiment",
@@ -47,8 +46,9 @@ __all__ = [
     "InputTag",
     "DatasetInput",
     "RunInputs",
-    "LiveSpan",
     "Span",
+    "LiveSpan",
+    "NoOpSpan",
     "SpanEvent",
     "SpanStatus",
     "SpanType",
