+++ mlflow/tracing/export/mlflow.py
@@ -10,7 +10,7 @@ from mlflow.entities import TraceData
 from mlflow.entities.trace_status import TraceStatus
 from mlflow.tracing.clients import TraceClient
 from mlflow.tracing.trace_manager import InMemoryTraceManager
-from mlflow.tracing.types.constant import (
+from mlflow.tracing.constant import (
     MAX_CHARS_IN_TRACE_INFO_METADATA_AND_TAGS,
     TRUNCATION_SUFFIX,
     SpanAttributeKey,
