+++ mlflow/tracing/fluent.py
@@ -9,13 +9,12 @@ from typing import Any, Callable, Dict, List, Optional
 from opentelemetry import trace as trace_api
 
 from mlflow import MlflowClient
-from mlflow.entities import LiveSpan, SpanType, Trace
+from mlflow.entities import LiveSpan, NoOpSpan, SpanType, Trace
 from mlflow.store.tracking import SEARCH_TRACES_DEFAULT_MAX_RESULTS
 from mlflow.tracing.display import get_display_handler
 from mlflow.tracing.provider import get_tracer
 from mlflow.tracing.trace_manager import InMemoryTraceManager
-from mlflow.tracing.types.constant import SpanAttributeKey
-from mlflow.tracing.types.wrapper import NoOpSpan
+from mlflow.tracing.constant import SpanAttributeKey
 from mlflow.tracing.utils import capture_function_input_args, encode_span_id
 from mlflow.utils import get_results_from_paginated_fn
