+++ tests/tracing/export/test_mlflow_exporter.py
@@ -4,7 +4,7 @@ from mlflow.entities import LiveSpan
 from mlflow.entities.trace_data import TraceData
 from mlflow.tracing.export.mlflow import MlflowSpanExporter
 from mlflow.tracing.trace_manager import InMemoryTraceManager
-from mlflow.tracing.types.constant import (
+from mlflow.tracing.constant import (
     MAX_CHARS_IN_TRACE_INFO_METADATA_AND_TAGS,
     TRUNCATION_SUFFIX,
     TraceMetadataKey,
